LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/transformations/2d.transformation.transform.nonfinite-expected.txt

2d.transformation.transform.nonfinite
transform() with Infinity/NaN is ignored
Actual output:
Expected output:


PASS transform() with Infinity/NaN is ignored
